chore(onboarding): absorb legacy steps into v2, delete legacy folder
The legacy /onboarding folder was orphaned routing-wise (PR #74 swapped
the route to OnboardingV2) but onboarding-v2 still imported three of
its files as a shared library. This PR pulls those into v2 and deletes
the legacy folder entirely.

- Add GENRES + SENTIMENT_RATINGS to onboarding-v2/data.js (single
  source for both constants).
- Replace the MoviesStepV2 thin wrapper with the full MoviesStep code
  (376 lines: TMDB live search, Supabase pool fetch, debounce, etc.).
  Default export renamed MoviesStep → MoviesStepV2 to match the file.
  GENRES import switched from ./GenresStep to ../data.
- Update SENTIMENT_RATINGS imports in OnboardingV2.jsx + RatingStepV2
  to source from local data.js.
- Update GENRES import in GenresStepV2 to source from local data.js.
- Move the two onboarding test files (OnboardingSteps,
  OnboardingSkip) into onboarding-v2/__tests__/ — they're stub-based
  re-implementations, portable as-is. Updated SENTIMENT_RATINGS
  import path inside.
- Delete src/features/onboarding/ — Onboarding.jsx (orphan main
  component), CinematicBackdrop.jsx (orphan), and the three step
  files now living in v2 alongside the rest.
- Drop the rollback comment in router.jsx — there's no legacy folder
  to roll back to anymore.
