+++ src/features/onboarding-v2/__tests__/OnboardingSkip.test.jsx
