+++ src/features/onboarding-v2/data.js
@@ -16,3 +16,36 @@ export const MAX_MOODS = 3
 
 // Local-storage key — consumed downstream by the home page if you wire it up.
 export const MOODS_LS_KEY = 'ff_onboarding_v2_moods'
+
+// === GENRES (Step 2) =======================================================
+// TMDb genre IDs paired with the dbName that matches movies.primary_genre in
+// Supabase. Used by GenresStepV2 (selection UI) + MoviesStepV2 (genre filter
+// for the Supabase pool fetch).
+export const GENRES = [
+  { id: 28,    name: 'Action',      dbName: 'Action'           },
+  { id: 12,    name: 'Adventure',   dbName: 'Adventure'        },
+  { id: 16,    name: 'Animation',   dbName: 'Animation'        },
+  { id: 35,    name: 'Comedy',      dbName: 'Comedy'           },
+  { id: 80,    name: 'Crime',       dbName: 'Crime'            },
+  { id: 99,    name: 'Documentary', dbName: 'Documentary'      },
+  { id: 18,    name: 'Drama',       dbName: 'Drama'            },
+  { id: 10751, name: 'Family',      dbName: 'Family'           },
+  { id: 14,    name: 'Fantasy',     dbName: 'Fantasy'          },
+  { id: 36,    name: 'History',     dbName: 'History'          },
+  { id: 27,    name: 'Horror',      dbName: 'Horror'           },
+  { id: 10402, name: 'Music',       dbName: 'Music'            },
+  { id: 9648,  name: 'Mystery',     dbName: 'Mystery'          },
+  { id: 10749, name: 'Romance',     dbName: 'Romance'          },
+  { id: 878,   name: 'Sci-Fi',      dbName: 'Science Fiction'  },
+  { id: 53,    name: 'Thriller',    dbName: 'Thriller'         },
+]
+
+// === RATINGS (Step 4) ======================================================
+// Sentiment → numeric rating mapping. Stored in user_ratings.rating during
+// completeOnboarding. The numeric scale lets later behavioral signals decay
+// or amplify these onboarding ratings consistently with in-app ratings.
+export const SENTIMENT_RATINGS = {
+  loved: 9,
+  liked: 7,
+  okay:  5,
+}
